Extract build specific logic to build crate, refactor crate
