+++ Cargo.toml
@@ -19,11 +19,9 @@ members = ["update_icons"]
 
 [dependencies]
 gtk = { version = "0.8", package = "gtk4" }
-
-[build-dependencies]
-gvdb = { version = "0.5.3", features = ["gresource"] }
-serde = { version = "1.0.197", features = ["derive"] }
-toml = "0.8.11"
+gvdb = { version = "0.5.3", features = ["gresource"], optional = true }
+serde = { version = "1.0.197", features = ["derive"], optional = true }
+toml = { version = "0.8.11", optional = true }
 
 [package.metadata.docs.rs]
 all-features = true
@@ -38,3 +36,8 @@ all = []
 # Enable entire icon kits
 icon-development-kit = []
 fluent-system-icons = []
+build-utils = [
+    "gvdb",
+    "serde",
+    "toml",
+]
