--- src/icon_names.rs
@@ -1,1 +1,0 @@
-include!(concat!(env!("OUT_DIR"), "/icon_names.rs"));
